+++ common.php
@@ -2,8 +2,8 @@
 
 /**
 *
-* @package UserReminder v1.4.0
-* @copyright (c) 2019, 2020 Mike-on-Tour
+* @package UserReminder v1.7.0
+* @copyright (c) 2019 - 2023 Mike-on-Tour
 * @license http://opensource.org/licenses/gpl-2.0.php GNU General Public License v2
 *
 */
@@ -92,9 +92,10 @@ class common
 	/**
 	* Remind users
 	*
-	* @param	array	$users_marked	Users selected for reminding identified by their user_id
+	* @param	array		$users_marked	Users selected for reminding identified by their user_id
+	*		boolean	$zeroposters	marks whether inactive users or zeroposters are to be handled, default is false to mark inactive users, set to true for handling zeroposters, necessary due to different config variables
 	**/
-	public function remind_users($users_marked)
+	public function remind_users($users_marked, $zeroposters = false)
 	{
 		if (count($users_marked) > 0)					// lets check for an empty array; just to be certain that none of the called functions throws an error or an exception
 		{
@@ -116,7 +117,8 @@ class common
 			*/
 			$this->email_arr = json_decode($this->config_text->get('mot_ur_email_texts'), true);
 			$now = time();
-			$reminder1 = $now - (self::SECS_PER_DAY * $this->config['mot_ur_days_reminded']);
+			// Since inactive users and zeroposters may have different time frames we have to distinguish here
+			$reminder1 = $zeroposters ? $now - (self::SECS_PER_DAY * $this->config['mot_ur_zp_days_reminded']) : $now - (self::SECS_PER_DAY * $this->config['mot_ur_days_reminded']);
 			// since we only have an array of user ids we need to get all the other user data from the DB and we start to select the users supposed to get the second reminder mail
 			// get only users we have selected before
 			// and who have been reminded once before
@@ -168,8 +170,8 @@ class common
 			}
 
 			//--------------------------------------------------------------------------------------
-			// and now we start to select the users supposed to get the first reminder mail
-			$day_limit = $now - (self::SECS_PER_DAY * $this->config['mot_ur_inactive_days']);
+			// and now we start to select the users supposed to get the first reminder mail, for this we have to calculate $day_limit depending on the type of user (inactive or zerooster)
+			$day_limit = $zeroposters ? $now - (self::SECS_PER_DAY * $this->config['mot_ur_zp_inactive_days']) : $now - (self::SECS_PER_DAY * $this->config['mot_ur_inactive_days']);
 			$query = 'SELECT user_id, username, user_email, mot_last_login, user_lang, user_timezone, user_dateformat, user_jabber, user_notify_type, mot_reminded_one, user_regdate, mot_sleeper_remind
 					FROM  ' . USERS_TABLE . '
 					WHERE ' . $this->db->sql_in_set('user_id', $users_marked) . '
@@ -204,10 +206,14 @@ class common
 
 				// all mails have been sent, let's set the reminder time(s)
 				$query = 'UPDATE ' . USERS_TABLE . ' SET mot_reminded_one = ' . $now;
-
-				if ($this->config['mot_ur_days_reminded'] == 0)		// if the admin selected to have only one reminder by setting this time frame to Zero ...
+				// if the admin selected to have only one reminder by setting this time frame to Zero we have to set this column too to enable deletion depending on the type of user
+				if ($zeroposters)
 				{
-					$query .= ', mot_reminded_two = ' . $now;		// ... we have to set this column too to enable deletion
+					$query .= $this->config['mot_ur_zp_days_reminded'] == 0 ? ', mot_reminded_two = ' . $now : '';
+				}
+				else
+				{
+					$query .= $this->config['mot_ur_days_reminded'] == 0 ? ', mot_reminded_two = ' . $now : '';
 				}
 
 				$query .= ' WHERE ' . $this->db->sql_in_set('user_id', $first_reminders_ary);